Add new EdgeFragment abstract type to reduce redundant code
