+++ server/src/graql/gremlin/fragment/InIsaFragment.java
@@ -57,7 +57,7 @@ import static grakn.core.server.kb.Schema.VertexProperty.LABEL_ID;
  */
 
 @AutoValue
-public abstract class InIsaFragment extends Fragment {
+public abstract class InIsaFragment extends EdgeFragment {
 
     @Override
     public abstract Variable end();
@@ -141,24 +141,7 @@ public abstract class InIsaFragment extends Fragment {
 
 
     @Override
-    public Set<Node> getNodes() {
-        Node start = new Node(NodeId.of(NodeId.NodeType.VAR, start()));
-        Node end = new Node(NodeId.of(NodeId.NodeType.VAR, end()));
-        Node middle = new Node(NodeId.of(NodeId.NodeType.ISA, new HashSet<>(Arrays.asList(start(), end()))));
-        middle.setInvalidStartingPoint();
-        return Sets.newHashSet(start, end, middle);
-    }
-
-    @Override
-    public Pair<Node, Node> getMiddleNodeDirectedEdge(Map<NodeId, Node> nodes) {
-        Node start = nodes.get(NodeId.of(NodeId.NodeType.VAR, start()));
-        Node middle = new Node(NodeId.of(NodeId.NodeType.ISA, new HashSet<>(Arrays.asList(start(), end()))));
-        // directed edge: middle -> start
-        return new Pair<>(middle, start);
-    }
-
-    @Override
-    public Set<Weighted<DirectedEdge>> directedEdges(Map<NodeId, Node> nodes) {
-        return directedEdges(NodeId.NodeType.ISA, nodes);
+    NodeId getMiddleNodeId() {
+        return NodeId.of(NodeId.NodeType.ISA, new HashSet<>(Arrays.asList(start(), end())));
     }
 }
