+++ server/src/graql/gremlin/fragment/InSubFragment.java
@@ -43,7 +43,7 @@ import java.util.Set;
  */
 
 @AutoValue
-public abstract class InSubFragment extends Fragment {
+public abstract class InSubFragment extends EdgeFragment {
 
     @Override
     public abstract Variable end();
@@ -86,16 +86,8 @@ public abstract class InSubFragment extends Fragment {
     }
 
     @Override
-    public Pair<Node, Node> getMiddleNodeDirectedEdge(Map<NodeId, Node> nodes) {
-        Node start = nodes.get(NodeId.of(NodeId.NodeType.VAR, start()));
-        Node middle = new Node(NodeId.of(NodeId.NodeType.SUB, new HashSet<>(Arrays.asList(start(), end()))));
-        // directed edge: middle -> start
-        return new Pair<>(middle, start);
-    }
-
-    @Override
-    public Set<Weighted<DirectedEdge>> directedEdges(Map<NodeId, Node> nodes) {
-        return directedEdges(NodeId.NodeType.SUB, nodes);
+    NodeId getMiddleNodeId() {
+        return NodeId.of(NodeId.NodeType.SUB, new HashSet<>(Arrays.asList(start(), end())));
     }
 }
